added branch - added folder pantiltweb_with_leds_button; added example
In this branch I added the functionality of a fifth webinterface button, a center-button, named 'Leds'. Pressing this button toggles the LEDs of a RGBW stick ON and Off.
I am working on an SPI communications interface to have a ANO encoder, connected to an Adafruit M4 Express, send encoder data via SPI to the Raspberry Pi 3B+ that has the PanTilt HAT mounted.